Add files via upload
This is a text-based version of the game "Chucklehead" (also known as Karma). The main file runs the program, the card file contains basic information for card classes, the hand file contains basic information for hand classes, the chuckleClasses file contains classes specific to Chucklehead, as well as the main bulk of the game being run. The commonGameFunctions file contains functions for common features used in games.